create a standalone grpc extension (#311)
Create a standalone extensions which just includes the kolide_grpc plugins for logger/config/distributed. 

The purpose of the standalone extension is to get Windows enrollment working without the runtime component, but can be useful for anyone who wants an alternative to the osquery TLS plugin for remote plugin functionality.