fix issue with duplicate log publish
updates the extension buffer logic to store a copy of the logID rather than the actual key.
when the View transaction is closed, there are no guarantees that the referenced `[]byte` key won't be written over,
which is what was happening by the time it got to the Delete call.