move go build targets to make.go (#385)
Moving the build targets to make.go further simplifies the bandaids that would otherwise be necessary to create builds on both unix and windows. This particular PR was driven by the desire to have clean outputs for binary names (launcher vs launcher.exe and extension.ext vs extension.exe) but the move to Go also allowed for consistency in various build targets.