Windows Auto Updater (#441)
Fix two bugs in the windows auto updated.

First, launcher finds the TUF repository by using fmt.Sprintf("%s-tuf", binaryName). Windows has a .exe appended to binaryName. Fix is a simple TrimSuffix.

Second, in an attempt to prevent corruption, launcher would not create the local TUF repository if the directory existed. Thus, when there is a problem creating a file, the local repository in an invalid state. Launcher assumes it’s good (directory exists), but TUF fails hard. This drops into a error loop. This is fixed by moving the isExists check from the directory to each file in the directory.