Update to windows service names (#457)
In https://github.com/kolide/launcher/pull/442 I updated the logic to strip out some characters. But, that’s in the wrong place. It needs to remove them from both the inferred `defaultName`, but also from a specified `ServiceName`. At least, probably… We might argue that if one specifies a bad service name, we shouldn’t correct it and instead through an error. However, correcting it here, allows the identifier to be set consistently across all platforms and modified in the narrow place it’s needed.