Fix windows service name appending (#462)
I noticed that we were getting a doubled SvcSvc on builds. This was clearly wrong.

Digging, it was because we always striped a trailing .exe and then appended a _svc. Thinking about it, I think it’s better to do that if, and only if. We’re setting a default name. An explicit ServiceName call should get what it expects. (After the character cleaning, and Camel Casing)