Postinstall script to generate an info file (#530)
We often want to understand what happened during installation. So the postinstall script should drop a config file onto disk with some basic info.

This is expected to be json, though on osx we additionally keep it as a plist.

To make this simpler, the postinstall templates have been extracted into the asset generation flow.