Fix windows osqueryd updates (#542)
Since the 0.11 autoupdate merge, windows osqueryd updates had not worked. This turns out to be because the osqueryd binary was specified as `osqueryd`, however it was actually `osqueryd.exe`. This mismatch caused the autoupdater routine and the findnewest function to use different paths.

autoupdate was downloading to `osqueryd-updates` while findnewest was looking in `osqueryd.exe-updates`

I think the underlying bug was because windows allows one to omit the `.exe` on executions. Which has the effect of masking an incorrect binary path.

This PR fixes this be ensure that the osqueryd path always has the trailing `.exe`. It also adds some additional logging.