Fix an infinit exec recursion bug
The FindNewestSelf function calls FindNewest. Which, in turns, calls `checkExecutable`. Which launchers the binary. Which goes back to the beginning. This bug only manifests for `launcher`, but not for the `osquery` side. This isn't a fatal error, as there's a 5s timeout on it. But it is slow.

Add an option to skip the final check, solving this. And add comments.