Fork Bomb Fixes (#601)
Fix inadvertent launcher fork bombs three ways. 

First, we add an environmental variable that _skips_ the initial re-exec functionality. This isn't documented, and is meant for internal use only. Frustratingly, we cannot just add it to flags, as we can't parse flags prior to exec.

Second, this risk primarily comes from the `FindNewestSelf` code path. This introduces `withRunningExectuable` as a mechanism to indicate that we should _not_ check a given path.

Third, if we can determine `os.Executable`, don't exec it. This is duplicative with the second option. But, it comes at it from a different direction. Seems better to have both