Allow set and override of osquery flags
Adds a new option `--osquery_flag` that allows the user to provide a
flag that will be passed directly to osquery. These flags are appended
to all other default and computed flags, allowing them to potentially
override the default and computed flags. This allows users to set flags
that would not otherwise be possible to use with Launcher (like
`windows_event_channels`), and also opens up additional uses of Launcher
with overridden options, plugins, etc.

Support is also added to the package tooling to build packages with this
flag set.

Closes #96