MSI User Interface (#625)
Add a basic user interface to the MSI. This is based on wix's WixUI_Minimal, but removes some of the screens we don't need. This is not currently exposed through package-builder.

Additionally change the MSI's ProductCode. This introduces a random factor, ensuring that each MSI has a unique ProductCode. Which means that any MSI will be able to install over other versions. This is an unusual choice from an MSI perspective, but it fit with out minimal usage of it