Additional error handling (and optional debug file logging) (#639)
Adds several additional logging capabilities. Panic is now caught, and logged. A handful of extra startup log lines have been added.

Adds the capability to mirror debug to a file. This greatly helps when debugging on windows, as extracting structed json from a file is simpler than extracting from the event log. This is something of an early preview -- I'd rather enable this by default, we until we have a log rotation ability, it won't work.

Lastly, this upgrades the go svc version