Set RunAtLoad to false in macOS launchd configuration
The existing behavior conflicted with the KeepAlive configuration that
only starts the daemon if the enroll secret path is present.

Setting this option to false leaves the general case the same: If a
package includes an enroll secret, the daemon is started immediately on
completion of install.

Setting this option to false allows the less-common case to work as
intended: The daemon will only start after an enroll secret is written
to the expected path.

The previous behavior would be to start the daemon even if no enroll
secret exists, and the daemon would then promptly exit with a nonzero status.