Add option to install Windows Service in disabled state
In some instances it is desirable to install the files and configure the
service but leave it in a disabled state until additional configuration
is performed.

This PR will be followed up with another that adds similar functionality
for systemd. There is no need to implement the same for macOS as the
launchd service is already configured to only start when the enroll
secret file exists.