Add Apple Notarization to packagekit (#687)
Add the apple notarization tools to launcher's packagekit. This open sources the backend end I've been using for notarization. Note that this is only a library, it is not plumbed through these command line tools.

This also adds some variable smuggling to context, as a way to pass data around during the notarization process.