Add a startup delay to the autoupdater,  to avoid interfering with critical work (#695)
When launcher first enrolls with a server, it often begins fetching key data. However, the autoupdater also starts. And if there's a pending update, this has the effect of pre-empting that early data.

This introduces a simple option to delay the autoupdater to allow key data to come in first. This option defaults to an hour.