Add macOS preinstall script to install rosetta2
As of the Big Sur general release, Apple M1 machines no longer ship
Rosetta 2. Instead, this must be installed manually. As osquery does
not yet ship a universal binary, launcher requires rosetta2.

During an interactive install, the user is prompted to okay a
rosetta2 install, but this does not appear to happen during an MDM
install. So, we need to trigger than from a preinstall script.

For internet discussion see:
- https://www.alansiu.net/2020/12/02/installing-rosetta-2-on-m1-apple-silicon-macs/
- https://derflounder.wordpress.com/2020/11/17/installing-rosetta-2-on-apple-silicon-macs/
- https://grahamgilbert.com/blog/2020/11/13/installing-rosetta-2-on-apple-silicon-macs/