Add macOS preinstall script to install rosetta2 (#708)
As of the Big Sur general release, Apple M1 machines no longer ship
Rosetta 2. Instead, this must be installed manually. As osquery does
not yet ship a universal binary, launcher requires rosetta2.

During an interactive install, the user is prompted to okay a
rosetta2 install, but this does not appear to happen during an MDM
install. So, we need to trigger than from a preinstall script.