Offset the updates
The osquery and launcher updaters run as seperate goroutines. If they
happen to trigger at the same time, it can get messy. One will restart,
before the other is done downloading, and this an update is missed.

The correct fix for this is to rearchtect away from this goroutine and
callback, but this might be a simple fix to reduce likelihood of errors.

This offsets one of the updaters, using the initial delay. As our
internvals are about an hour, this offset should reduce the likelihood
of this happening.