Don't always use cgo
For reasons unclear to me, I cannot trivially build `net/http` with CGO.
This is true even with a trivial example that imports `_ "net/http"`.
Reading a bit, it appear the `net/http` needs some kind of static flags,
and I'm not sure how to evaluate the effects of those, _so_ stop
building osquery-extension with cgo, it doesn't need it.

This change is somewhat larger than expected -- it shifts the way
make.go invokes targets, and allows a somewhat finer grained control.
It's not perfect, but it's some progress. And I don't want to do a
deeper refactor.