Rename the initial kolide plugin to `kolide_grpc`
Testing with v0.12.0, I'm running into some issues. I'm pretty sure they're because the extension registers as `kolide_initial`, but various things are configured to use `kolide_grpc`. (Which is kinda :shrug: since there are sub extensions with that name. But who knows)

Renaming this to kolide_grpc feels like the quickest way. It seems to work in some testing, but I want to get it a bit wider