Change flag handling to flag.ContinueOnError
If we remove a config option, we cause launcher to not startup. This
is a bit scary. One approach would be to change the flag handling to
`flag.ContinueOnError`.

This has the advantage of letting things just work.

But the downside, is that `launcher --garbage-flag` won't prouce an error.

I'm honestly on the fence about it.