Remove `NonProdOnly` and rename `IsSeparator` to `Separator`
Working a bit with the generator side of this, I think we don't need
NonProdOnly. It was a toss up between not implementing it in the
generator, or removing it. I opted to remove it, on the theory that
less code is less bugs.

I opted to rename `IsSeparator` to simply `Separator`. Half to avoid
snake case vs camel case silliness. And half because it seemed simpler.